move demo model class to common lib
