+++ demo/lightning/text-based/__init__.py
@@ -1,0 +1,1 @@
+from .demo_model import *
